CloudProviders/gcp/examples/cloud-functions/requirements.txt
google-auth==1.24.0
google-cloud-firestore>=2.0.2
google-cloud-pubsub>=2.2.0
google-cloud-secret-manager>=2.1.0
google-cloud-storage>=1.35.0
hypothesis==6.0.1
hypothesis-jsonschema==0.19.0
idna==2.10
jsonschema==3.2.0
pyrsistent==0.17.3
python-dotenv==0.15.0
pytest-cov>=2.11.1
pytest-mock>=3.5.1
PyYAML==5.3.1
requests>=2.25.1
urllib3==1.26.2
